:recycle: move ld_r_n
