+++ src/__test__/instructions/ld_r_n/ld_c_n.test.ts
@@ -1,9 +1,9 @@
 import { CpuOperation } from "vm/cpu_operation";
-import { LdDN } from "vm/instructions/ld_d_n";
+import { LdCN } from "vm/instructions/ld_r_n/ld_c_n";
 import { Memory } from "vm/memory";
 import { RegisterSet } from "vm/register/register_set";
 
-describe("LD D, n test", () => {
+describe("LD C, n test", () => {
   let buffer: ArrayBuffer;
   let register: RegisterSet;
 
@@ -26,20 +26,20 @@ describe("LD D, n test", () => {
     register.SP = 0x99AA;
     const prevRegister = register.clone();
 
-    const ld = new LdDN(new CpuOperation(register, new Memory(new Uint8Array(buffer))));
+    const ld = new LdCN(new CpuOperation(register, new Memory(new Uint8Array(buffer))));
     const cycle = ld.exec();
 
     // 返値(サイクル数)の確認
     expect(cycle).toBe(8);
     // 他のレジスタに影響を与えていないことの確認
     expect(register.AF).toBe(prevRegister.AF);
-    expect(register.BC).toBe(prevRegister.BC);
+    expect(register.DE).toBe(prevRegister.DE);
     expect(register.HL).toBe(prevRegister.HL);
     expect(register.SP).toBe(prevRegister.SP);
     // 指定のレジスタの値が変わっていることの確認
-    expect(register.DE).toBe(0xAB66);
-    expect(register.D).toBe(0xAB);
-    expect(register.E).toBe(0x66);
+    expect(register.BC).toBe(0x33AB);
+    expect(register.B).toBe(0x33);
+    expect(register.C).toBe(0xAB);
     // プログラムカウンタが進んでいることの確認
     expect(register.PC).toBe(prevRegister.PC + 2);
   });
